Rename replicated rdb file only if it is valid. (#1)
If leader has a memory corruption and managed to replicate a corrupted dump.rdb to all of it's followers, there is a chance for data loss. The follower renames temporary dump.rdb file before doing a rdbLoad and checking the validity of newly replicated file. Fixed this by checking validity of newly replicated dump.rdb file before attempting to rename the temporary file.